Move the unit and application display status functions to the core/status package.
Both the functions and the tests are moved into the core package.
Nothing else is touched here.
